Fixed Histogram: halved memory footprint, test boundary conditions, dropped unnecessary binary search.
These changes preserve/correct behavior of the current Histogram.

src/java/voldemort/store/stats/Histogram.java
- Halved memory footprint by dropping unnecessary "bounds" array
- Dropped unnecessary binary search, making insert O(1) rather than O(log(nBuckets))
- Improved documentation
- Made interface consistent for type of values inserted/got from histogram (i.e., all are 'long')

test/unit/voldemort/store/stats/HistogramTest.java
- Added tests for boundary conditions: -ive values are dropped on
  insert, "too large" values are bucketed in the final bucket on
  insert.

Minor fixes to RequestCounter calls to histogram to conform to 'long' interfaces.

+++ src/java/voldemort/store/stats/RequestCounter.java
@@ -17,8 +17,8 @@ public class RequestCounter {
     private final int durationMS;
     private final Time time;
     private final Histogram histogram;
-    private volatile int q95LatencyMs;
-    private volatile int q99LatencyMs;
+    private volatile long q95LatencyMs;
+    private volatile long q99LatencyMs;
     private boolean useHistogram;
 
     /**
@@ -233,11 +233,11 @@ public class RequestCounter {
         return getValidAccumulator().getAllMaxCount;
     }
 
-    public int getQ95LatencyMs() {
+    public long getQ95LatencyMs() {
         return q95LatencyMs;
     }
 
-    public int getQ99LatencyMs() {
+    public long getQ99LatencyMs() {
         return q99LatencyMs;
     }
